requirements.txt
networkx>=3.0
pydantic>=2.0
google-genai>=0.6.0
requests
pyvis>=0.3.2
python-dotenv
